Fourth commit.
Changed Deck so that instead of being a class to be instantiated it now
just holds methods to generate ArrayLists of Card objects, as I
recognized that was all I wanted and needed as my deck of cards. Also
began to fill in BlackjackEngine with methods to initialize the needed
decks, one as a reference and the other as the deck cards will be
removed from, retrieve numeric values for cards based on the rules of
blackjack, and deal cards to the computer dealer and human players. A
turn method has been created but still needs to be filled with win/lose
conditions and player options.