Fifth commit.
The blackjack class has now achieved minimum functionality, after reworking code that was being copied and pasted into separate helper methods, renaming the turn method to the round method and filling in the body, and adding a game method that handles the operation of the entire game from initialization to completion, in which the round is the primary interactable component. Whether players win/lose is indicated by printed statements but not explicitely stated nor tracked within the code.