Sixth commit.
Fixed an error relating to when the dealer held and created a new
evaluate method to declare which players won, tied, or lost to the
dealer's hand. In addition, much of the code was refactored to print the
cards the dealer/players drew and their total scores before each
draw or hold decision, the gameDone variable was removed once it became
redundant, and numerous smaller errors were fixed. Besides further
testing, hiding the dealer's second card and resulting score until their
turn still needs to be implemented.