Seventh commit.
Implemented the dealer hiding their second card and reworked most of the
methods to more accurately follow the rules of blackjack. Due to
repeating code in two separate places, a dealerReveal method was created
to handle when and why the dealer would reveal their second card and the
subseqent impact on the game. The evaluateEnd method was expanded to
account for the namesake of the game, the occurence of a blackjack which
automatically ends the round should the dealer get it and grants any
player who gets it an automatic win unless the dealer also gets a
blackjack, in which case a tie occurs. At this point in time I am more
certain that the fundamental rules have been accounted for, and what
remains in the future is handling of 'bets,' splitting pairs, and
eventually functioning graphics.